feat(register): 修改了注册的逻辑

+++ src/main.rs
@@ -81,15 +81,16 @@ impl<T: Serialize> ResJson<T> {
 /// 通用字段
 #[derive(Serialize, Deserialize, Debug, Clone, Copy)]
 struct UniversalField {
-    id: usize,
+    /// ID为数据库自增，不会进入迭代器
+    id: u64,
     /// 乐观锁
-    version: usize,
+    version: u32,
     /// 创建人
-    create_by: Option<usize>,
+    create_by: Option<u64>,
     /// 创建时间
     create_time: Option<DateTime<Utc>>,
     /// 更新人
-    update_by: Option<usize>,
+    update_by: Option<u64>,
     /// 更新时间
     update_time: Option<DateTime<Utc>>,
     /// 是否删除，默认为false
